clean usless and arrange
