+++ blockchain/api/errors.go
@@ -18,8 +18,5 @@ package api
 
 import "errors"
 
-var ErrNilBlock = errors.New("block is nil")
-var ErrSyncProcessing = errors.New("Sync is in progress")
-var ErrGetLastBlock = errors.New("failed get last block")
 var ErrCreateGenesisBlock = errors.New("Error in creating Genesis block")
 var ErrCommitBlock = errors.New("Error in committing block")
