--- blockchain/grpc_command_service.go
@@ -1,23 +1,0 @@
-/*
- * Copyright 2018 It-chain
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package blockchain
-
-type GrpcCommandService interface {
-	RequestBlock(peerId PeerId, height uint64) error
-	ResponseBlock(peerId PeerId, block Block) error
-	SyncCheckResponse(block Block) error
-}
